requirements.txt
django >= 3.2.8, < 4.0
django-admin >= 2.0.1
django-jinja >= 2.9, < 3.0
mysqlclient >= 2.0.3, < 3.0
pymemcache >= 3.5.0, < 4.0
python-slugify >= 5.0.2, < 6.0
gunicorn >= 20.1.0, < 21.0
